requirements.txt
Adafruit-Blinka==8.56.0
adafruit-circuitpython-busdevice==5.2.11
adafruit-circuitpython-connectionmanager==3.1.3
adafruit-circuitpython-framebuf==1.6.7
adafruit-circuitpython-requests==4.1.10
adafruit-circuitpython-ssd1306==2.12.19
adafruit-circuitpython-typing==1.11.2
Adafruit-PlatformDetect==3.77.0
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Cython==3.0.12
daemonize==2.5.0
Flask==3.1.0
flask-swagger-ui==4.11.1
future==1.0.0
gpiod==2.3.0
idna==3.10
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpickle==4.0.5
MarkupSafe==3.0.2
packaging==24.2
pillow==11.1.0
pyftdi==0.56.0
pyproject_hooks==1.2.0
pyserial==3.5
pyusb==1.3.1
PyYAML==6.0.2
# sudo env/bin/pip install ../reedsolomon/dist/reedsolo-2.1.2b1-cp312-cp312-linux_armv7l.whl
#reedsolo @ file:///home/chip/reedsolomon/dist/reedsolo-2.1.2b1-cp312-cp312-linux_armv7l.whl#sha256=f05e2a5616e71cf3d146250ad66edfbdf783b74c7ceb6d1fcd992c51140ab868
requests==2.32.3
smbus2==0.5.0
sysv_ipc==1.1.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3